Monolight_Latest/requirement.txt.txt
cmake
pillow
opencv-python
dlib
numpy
pandas
tensorflow